+++ packages/shield-controller/src/backend.ts
@@ -2,6 +2,8 @@ import type { SignatureRequest } from '@metamask/signature-controller';
 import type { TransactionMeta } from '@metamask/transaction-controller';
 
 import type {
+  CheckCoverageRequest,
+  CheckSignatureCoverageRequest,
   CoverageResult,
   CoverageStatus,
   LogSignatureRequest,
@@ -40,6 +42,8 @@ export type GetCoverageResultRequest = {
 };
 
 export type GetCoverageResultResponse = {
+  message?: string;
+  reasonCode?: string;
   status: CoverageStatus;
 };
 
@@ -74,30 +78,50 @@ export class ShieldRemoteBackend implements ShieldBackend {
     this.#fetch = fetchFn;
   }
 
-  async checkCoverage(txMeta: TransactionMeta): Promise<CoverageResult> {
-    const reqBody = makeInitCoverageCheckBody(txMeta);
+  async checkCoverage(req: CheckCoverageRequest): Promise<CoverageResult> {
+    let { coverageId } = req;
+    if (!coverageId) {
+      const reqBody = makeInitCoverageCheckBody(req.txMeta);
+      ({ coverageId } = await this.#initCoverageCheck(
+        'v1/transaction/coverage/init',
+        reqBody,
+      ));
+    }
 
-    const { coverageId } = await this.#initCoverageCheck(
-      'v1/transaction/coverage/init',
-      reqBody,
-    );
-
-    const coverageResult = await this.#getCoverageResult(coverageId);
-    return { coverageId, status: coverageResult.status };
+    const txCoverageResultUrl = `${this.#baseUrl}/v1/transaction/coverage/result`;
+    const coverageResult = await this.#getCoverageResult(coverageId, {
+      coverageResultUrl: txCoverageResultUrl,
+    });
+    return {
+      coverageId,
+      message: coverageResult.message,
+      reasonCode: coverageResult.reasonCode,
+      status: coverageResult.status,
+    };
   }
 
   async checkSignatureCoverage(
-    signatureRequest: SignatureRequest,
+    req: CheckSignatureCoverageRequest,
   ): Promise<CoverageResult> {
-    const reqBody = makeInitSignatureCoverageCheckBody(signatureRequest);
+    let { coverageId } = req;
+    if (!coverageId) {
+      const reqBody = makeInitSignatureCoverageCheckBody(req.signatureRequest);
+      ({ coverageId } = await this.#initCoverageCheck(
+        'v1/signature/coverage/init',
+        reqBody,
+      ));
+    }
 
-    const { coverageId } = await this.#initCoverageCheck(
-      'v1/signature/coverage/init',
-      reqBody,
-    );
-
-    const coverageResult = await this.#getCoverageResult(coverageId);
-    return { coverageId, status: coverageResult.status };
+    const signatureCoverageResultUrl = `${this.#baseUrl}/v1/signature/coverage/result`;
+    const coverageResult = await this.#getCoverageResult(coverageId, {
+      coverageResultUrl: signatureCoverageResultUrl,
+    });
+    return {
+      coverageId,
+      message: coverageResult.message,
+      reasonCode: coverageResult.reasonCode,
+      status: coverageResult.status,
+    };
   }
 
   async logSignature(req: LogSignatureRequest): Promise<void> {
@@ -159,12 +183,19 @@ export class ShieldRemoteBackend implements ShieldBackend {
 
   async #getCoverageResult(
     coverageId: string,
-    timeout: number = this.#getCoverageResultTimeout,
-    pollInterval: number = this.#getCoverageResultPollInterval,
+    configs: {
+      coverageResultUrl: string;
+      timeout?: number;
+      pollInterval?: number;
+    },
   ): Promise<GetCoverageResultResponse> {
     const reqBody: GetCoverageResultRequest = {
       coverageId,
     };
+
+    const timeout = configs?.timeout ?? this.#getCoverageResultTimeout;
+    const pollInterval =
+      configs?.pollInterval ?? this.#getCoverageResultPollInterval;
 
     const headers = await this.#createHeaders();
     return await new Promise((resolve, reject) => {
@@ -179,14 +210,11 @@ export class ShieldRemoteBackend implements ShieldBackend {
         // eslint-disable-next-line no-unmodified-loop-condition
         while (!timeoutReached) {
           const startTime = Date.now();
-          const res = await this.#fetch(
-            `${this.#baseUrl}/v1/transaction/coverage/result`,
-            {
-              method: 'POST',
-              headers,
-              body: JSON.stringify(reqBody),
-            },
-          );
+          const res = await this.#fetch(configs.coverageResultUrl, {
+            method: 'POST',
+            headers,
+            body: JSON.stringify(reqBody),
+          });
           if (res.status === 200) {
             return (await res.json()) as GetCoverageResultResponse;
           }